main-20250407T134144Z-001/main/requirements.txt
xmltodict
pyqrcode
pypng
pypdf
rlPyCairo
reportlab